+++ src/mask.rs
@@ -1,5 +1,10 @@
-use core::fmt;
-use derive_more::{BitAnd, BitAndAssign, BitOr, BitOrAssign, Deref, From, Not};
+use core::{
+    fmt,
+    ops::{Shl, ShlAssign, Shr, ShrAssign},
+};
+use derive_more::{
+    BitAnd, BitAndAssign, BitOr, BitOrAssign, BitXor, BitXorAssign, Deref, From, Not,
+};
 
 /// A struct representing a mask of 16 bits, used for Ethereum trie operations.
 ///
@@ -17,10 +22,12 @@ use derive_more::{BitAnd, BitAndAssign, BitOr, BitOrAssign, Deref, From, Not};
     Ord,
     Deref,
     From,
-    BitAnd,
     BitAndAssign,
+    BitAnd,
     BitOr,
     BitOrAssign,
+    BitXor,
+    BitXorAssign,
     Not,
 )]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
@@ -34,6 +41,9 @@ impl fmt::Debug for TrieMask {
 }
 
 impl TrieMask {
+    /// The size of this mask in bits.
+    pub const BITS: u32 = u16::BITS;
+
     /// Creates a new `TrieMask` from the given inner value.
     #[inline]
     pub const fn new(inner: u16) -> Self {
@@ -92,5 +102,49 @@ impl TrieMask {
     #[inline]
     pub const fn unset_bit(&mut self, index: u8) {
         self.0 &= !(1u16 << index);
+    }
+}
+
+impl<T> Shl<T> for TrieMask
+where
+    u16: Shl<T, Output = u16>,
+{
+    type Output = Self;
+
+    #[inline]
+    fn shl(self, rhs: T) -> Self::Output {
+        Self(self.0.shl(rhs))
+    }
+}
+
+impl<T> ShlAssign<T> for TrieMask
+where
+    u16: ShlAssign<T>,
+{
+    #[inline]
+    fn shl_assign(&mut self, rhs: T) {
+        self.0.shl_assign(rhs);
+    }
+}
+
+impl<T> Shr<T> for TrieMask
+where
+    u16: Shr<T, Output = u16>,
+{
+    type Output = Self;
+
+    #[inline]
+    fn shr(self, rhs: T) -> Self::Output {
+        Self(self.0.shr(rhs))
+    }
+}
+
+impl<T> ShrAssign<T> for TrieMask
+where
+    u16: ShrAssign<T>,
+{
+    #[inline]
+    fn shr_assign(&mut self, rhs: T) {
+        self.0.shr_assign(rhs);
     }
 }