+++ CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.3](https://github.com/alloy-rs/trie/releases/tag/v0.9.3) - 2026-01-07
+
+### Features
+
+- Add bit ops to TrieMask ([#117](https://github.com/alloy-rs/trie/issues/117))
+
+## [0.9.2](https://github.com/alloy-rs/trie/releases/tag/v0.9.2) - 2025-12-22
+
+### Features
+
+- Add ordered_trie_root_encoded for pre-encoded items ([#115](https://github.com/alloy-rs/trie/issues/115))
+
+### Miscellaneous Tasks
+
+- Release 0.9.2
+- Re-use alloy-primitives keccak empty ([#113](https://github.com/alloy-rs/trie/issues/113))
+- `missing-const-for-fn` lint back to "warn". ([#112](https://github.com/alloy-rs/trie/issues/112))
+
 ## [0.9.1](https://github.com/alloy-rs/trie/releases/tag/v0.9.1) - 2025-08-19
 
 ### Documentation
@@ -14,6 +32,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Features
 
 - Retain proofs of non-target nodes in certain edge-cases. ([#109](https://github.com/alloy-rs/trie/issues/109))
+
+### Miscellaneous Tasks
+
+- Release 0.9.1
 
 ## [0.9.0](https://github.com/alloy-rs/trie/releases/tag/v0.9.0) - 2025-06-20